Add util function for getting Elasticsearch instance
